CARD_management

Manage your different sets of variables to extract. For CARD advanced users, this function manages the CARD directory structure by performing automatic file operations to get the CARD parameterization files to your wanted directory in order to custom them or create yours to use them in a second step with CARD_extraction(). If you want to submit request for new CARD see the GitHub repohttps://github.com/super-lou/CARD.

Usage

CARD_management(
  CARD_name = c("QA", "QJXA"),
  CARD_path = "./WIP",
  add_id = TRUE,
  overwrite = FALSE,
  verbose = FALSE
)

Arguments

CARD_name

A vector of character strings to specify which variables you want to extract. See CARD_list_all() to get the variable names. By default, c("QA", "QJXA"). If NULL, all the variable will be extracted, so avoid this value except with extract_only_metadata = TRUE.

CARD_path

An optional character string for the path where to search for custom CARDs that have been created by the CARD_management function. By default, NULL in order to get the default CARD variable parameters.

add_id

logical. If TRUE, numerical IDs will be added to the start of the copied and pasted CARD names to maintain the input order. Default is TRUE.

overwrite

logical. If TRUE, existing CARD files in the analysis directory will be overwritten. Default is TRUE.

verbose

logical. Should intermediate messages be printed during the execution of the function ? Default FALSE.

Value

Selected CARD parameterization files will be created in the CARD_path directory.

See also

Examples

# Get the QA and QMNA CARD variables in your local CARD_path directory
CARD_management(CARD_name=c("QA", "QMNA"),
                CARD_path="CARD-WIP",
                overwrite = TRUE)