Model Building

Simplify the process of building multiple models in a sequential order. This is particularly helpful in epidemiological cases of testing effect of additional parameters. Every parameter should be theoretically a part of the causal model for the exposure-outcome relationship.

build_sequential_models(formula, data, exposure = NULL, engine = "lm")

Arguments

formula
an object of class formula that shows the names of the outcomes (can be more than 1) and the names of the predictors (which should contain the exposure variable).

data
data frame or data table (or tibble) that contains the named variables

exposure
Variable that is forced to be maintained in every model as a predictor.

engine
Set the "engine" or the regression tool that will be used

Value

A tidy tibble of models. Each one will likely be grouped by its outcome, and then with sequential columns using increased/additive models. Each model, in a tidy format, will have two additional columns.

outcomes identifies which outcome was used for the specific regression

covar number of covariates used in sequence of predictors given, with exposure always being placed in position 1

Details

This is considering what is available with the modelr package and the tidymodels approach, and finding an in-between for the causality / epidemiology approach of building intentional, sequentional models. Expect changes in the process, and potential future dependencies on the tidymodels appraoches.

Examples

data(geh)
f <- svg_mag + qrs_tang ~ lab_hba1c + bmi
build_sequential_models(f, data = geh)

#> # A tibble: 10 x 9
#>    outcomes term  estimate std.error statistic  p.value conf.low conf.high covar
#>    <chr>    <chr>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <dbl>
#>  1 svg_mag  (Int…  4.20e+3   243.      17.3    1.89e-52 3724.     4678.        1
#>  2 svg_mag  lab_… -5.10e+1    29.0     -1.76   7.96e- 2 -108.        6.04      1
#>  3 qrs_tang (Int…  4.84e+1     8.02     6.04   3.12e- 9   32.7      64.2       1
#>  4 qrs_tang lab_…  2.61e-1     0.958    0.273  7.85e- 1   -1.62      2.14      1
#>  5 svg_mag  (Int…  4.19e+3   386.      10.9    1.11e-24 3432.     4950.        2
#>  6 svg_mag  lab_… -5.09e+1    29.1     -1.75   8.02e- 2 -108.        6.15      2
#>  7 svg_mag  bmi    3.45e-1    10.4      0.0332 9.74e- 1  -20.0      20.7       2
#>  8 qrs_tang (Int…  4.76e+1    12.8      3.73   2.16e- 4   22.5      72.6       2
#>  9 qrs_tang lab_…  2.63e-1     0.960    0.275  7.84e- 1   -1.62      2.15      2
#> 10 qrs_tang bmi    3.05e-2     0.343    0.0890 9.29e- 1   -0.643     0.704     2