General Interface for Circular Regression Models

circular_reg() is a parsnip friendly method for specification of circular regression model before fitting. When using the circular::lm.circular engine, a formula is not given/needed.

circular_reg(
  mode = "regression",
  pattern = NULL,
  initial = NULL,
  tolerance = NULL
)

# S3 method for circular_reg
update(
  object,
  pattern = NULL,
  initial = NULL,
  tolerance = NULL,
  fresh = FALSE,
  ...
)

# S3 method for circular_reg
print(x, ...)

Arguments

mode
A character string that describes the type of model. In this case, it only supports type of "regression", which is default.

pattern
A character string of either "c-c" or "c-l" which show the relationship between the dependent and independent variables, identifying if they are circular or not. This changes the additional parameters available.

If "c-c" is selected, neither initial or tolerance are required

If "c-l" is selected, both initial or tolerance are required

initial
A vector with initial values the length equal to the columns of the independent variable, plus the intercept. For example, with 3 predictors (and one intercept), the initial value should be x = rep(0, 4)

tolerance
A numerical value, which defaults to 1e-10, can be set at a lower or higher tolerance which sets the accuracy for algorithm convergence.

object
Circular model specification

fresh
A logical for whether the arguments should be modified in place or replaced altogether

...
Extensible

x
circular model specification

Examples

library(magrittr)
library(circular)

#> 
#> Attaching package: ‘circular’

#> The following objects are masked from ‘package:stats’:
#> 
#>     sd, var

library(parsnip)
f <- az_svg ~ lab_hba1c + cad
df <- geh[c("az_svg", "lab_hba1c", "cad")]
df$az_svg <- circular(df$az_svg, units = "degrees") %>% conversion.circular(., units = "radians")
circular_reg(pattern = "c-l", initial = rep(0, 3), tolerance = 1e-3) %>%
  set_engine("circular") %>%
  fit(f, data = df)

#> Warning: an object is coerced to the class 'circular' using default value for the following components:
#>   type: 'angles'
#>   units: 'radians'
#>   template: 'none'
#>   modulo: 'asis'
#>   zero: 0
#>   rotation: 'counter'
#> conversion.circularyradians0counter2pi

#> Iteration  1 :    Log-Likelihood =  544.6186 
#> Iteration  2 :    Log-Likelihood =  544.6203 
#> Iteration  3 :    Log-Likelihood =  544.6204 

#> parsnip model object
#> 
#> Fit time:  21ms 
#> 
#> Call:
#> lm.circular.cl(y = ..2, x = ..1, init = ..3, verbose = TRUE,     tol = ..4)
#> 
#> 
#>  Circular-Linear Regression 
#> 
#>  Coefficients:
#>       Estimate Std. Error t value Pr(>|t|)
#> [1,] -0.045124   0.048211   0.936    0.175
#> [2,]  0.005829   0.005731   1.017    0.155
#> [3,] -0.019320   0.034200   0.565    0.286
#> 
#>  Log-Likelihood:  544.6 
#> 
#>  Summary: (mu in radians)
#>   mu:  0.1342 ( 0.02233 )  kappa:  4.731 ( 0.2814 )
#> p-values are approximated using normal distribution
#>