KGE

Computes the Kling-Gupta efficiency coefficient. This function was largely inspired by the similar function in the HydroGOF package. Here, only a simplified version is provided with no checks on inputs and very little formatting to speed it up.

Usage

compute_KGE(obs, sim, na.rm = TRUE, method = 1)

Arguments

obs

Observed streamflow vector

sim

Simulated streamflow vector

na.rm

Should missing values be omited?

method

Two methods are implemented (see hydroGOF::KGE): '1' for Gupta et al. (2009) and '2' for Kling et al. (2012)

Value

KGE