Center for Medicare and Medicaid Services (CMS) Procedure Codes

This is a generative function used to call CMS procedure codes. It is used to create a dataset that can be generally used to map procedure codes to their descriptions, allowing for understanding of interventions performed. The currently supported codes are explained in the details.

The following procedure codes are currently supported:

ICD9 procedure codes, most recently updated on 2014-10-01

ICD10 procedure codes, most recently updated on 2023-01-11

HCPCS prcoedure codes, most recently updated on 2023-11-29

CPT procedure codes, most recently updated on 2023-11-29

Usage

procedure_codes(format, version)

Arguments

format

The format of the procedure codes, written as a character. Currently supported formats are: c("icd9", "icd10", "hcpcs", "cpt") (case-insensitive).

version

The version of the procedure codes, which are written as a character although the majority are listed as a year. Currently supported: c("2014", "2023")

Value

A tbl_df with two columns: code and description. The code refers to the procedure code, while the description refers to the description of the procedure.

Details

CMS will usually release updated version of these codes on an annual basis. Each dataset that is supported below can be identified by the year it was published (not the go-live date, but the publically-available date). The previous versions that are included in the package are as below.

ICD9: 2014

ICD10: 2023

HCPCS: 2023

CPT: 2023