Recurrent Event Summary Table by Group

recur_summary Creates a table with summary of recurrent events

recur_summary(data, covar)

Arguments

data
Recurrent event data in marginal format. There must be an ID column. Must merge in the covariate of interest into this data set.

covar
Name of covariate of interest to serve as grouping variable.

Value

Summary table by grouping variable, can be placed into a latex environment with kableExtra. Assumes that death events may be present when most recent non-EVENT has status 1.

Details

This function allows for taking the output of card::recur_survival_table marginal format repeat event data, and creates a summary table that describes the number of events by strata/event.