Recurrent Event Sequential Model Building

Takes a different covariate groups to generate several models for recurrent event survival analyses.

recurrent_model_building(data, covar.builds, model.type, prop.scores = NULL)

Arguments

data
Data frame that is the survival format, potentially made by the card::recur_survival_table. Has to be merged with the superset of covariates that are being tested.

covar.builds
This is a vector that names the individual vectors for each model, likely sequential and additive. The individual vectors contain the names of the columns in the data frame that will generate regressions.

model.type
Type of recurrent event data, selected from c("marginal", "pwptt", "pwpgt")

prop.scores
This is a vector of the names of which covar.builds should be performed with propensity weighting. This will call a separate function card::recurrent_propensity that will generate both a PROP_SCORE column and PROP_WEIGHT column. Optional parameter, defaults to NULL.

Value

List of models in sequential order.

Details

Using the survival models in different types (e.g. marginal, PWP, etc), to create Cox regressions that are in a sequential order. Using the covariates given, will create the models on the fly. Need to specify model type and provide data in a certain format.