Tidy a(n) cosinor object

Tidy summarizes information about the components of a cosinor model.

Usage

# S3 method for class 'cosinor'
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x

A cosinor object created by cosinor()

conf.int

Logical indicating whether or not to include confidence

interval in tidied output

conf.level

The confidence level to use if conf.int = TRUE. Must be

between 0 and 1, with default to 0.95 (the 95% confidence interval).

...

For extensibility

Value

a tibble object

Details

cosinor objects do not necessarily have a T-statistic as the standard error is not based on a mean value, but form a joint-confidence interval. The standard error is generated using Taylor series expansion as the object is a subspecies of harmonic regressions.