Multiple Component Cosinor Features

Extract the special/global features of a multiple component cosinor. In a multiple component model, there are specific parameters that are not within the model itself, but must be extracted from the model fit. When extracted, can be used to improve the plot of a multiple component cosinor. However, this is only possible if the cosinor is harmonic (see details). For single-component models, the orthophase is the same as the acrophase and the global amplitude

Global Amplitude (Ag) = the overall amplitude is defined as half the difference between the peak and trough values

Orthophase (Po) = the lag until the peak time

Bathyphase (Pb) = the lag until the trough time

cosinor_features(object, population = TRUE, ...)

Arguments

object
Model of class cosinor with multiple periods

population
If the object is a population cosinor, should the features be calculated for the individual cosinors or for the population-cosinors. Default is TRUE. This has no effect on "Individual" cosinor objects.

If TRUE, then will calculate features for entire population.

If FALSE, then will calculate features for every individual cosinor in the population.

...
For extensibility

Value

When returning the cosinor features for a single model, will return an object of class list. When returning the cosinor features for every individual in a population cosinor, will return an object of class tibble.

Details

These calculations can only occur if the periods of the cosinor are harmonic - as in, the longest period is a integer multiple of the smallest period (known as the fundamental frequency). Otherwise, these statistics are not accurate or interpretable.

Examples

data(twins)
model <- cosinor(rDYX ~ hour, twins, c(24, 8), "patid")

#> 23 subjects were removed due to having insufficient observations.

results <- cosinor_features(model, population = FALSE)

#> This is a harmonic multiple-component cosinor object. The orthophase, bathyphase, and global amplitude were calculated.

head(results)

#> # A tibble: 6 x 7
#>   population harmonic  peak trough ampGlobal orthophase bathyphase
#>        <dbl> <lgl>    <dbl>  <dbl>     <dbl>      <dbl>      <dbl>
#> 1          1 TRUE      3.35   1.81     0.768      23          12  
#> 2          2 TRUE      3.05   2.25     0.398       3.71       15.9
#> 3          7 TRUE      2.96   2.30     0.330      10.4        23.0
#> 4          8 TRUE      4.23   2.41     0.909      13           1  
#> 5          9 TRUE      2.64   2.19     0.222       9          21  
#> 6         10 TRUE      3.85   2.50     0.671      14           2