Multiple Component Cosinor Features

Extract the special/global features of a multiple component cosinor. In a multiple component model, there are specific parameters that are not within the model itself, but must be extracted from the model fit. When extracted, can be used to improve the plot of a multiple component cosinor. However, this is only possible if the cosinor is harmonic (see details). For single-component models, the orthophase is the same as the acrophase and the global amplitude

Global Amplitude (Ag) = the overall amplitude is defined as half the difference between the peak and trough values

Orthophase (Po) = the lag until the peak time

Bathyphase (Pb) = the lag until the trough time

cosinor_features(object, ...)

Arguments

object
Model of class cosinor with multiple periods

...
For extensibility

Value

List of features

Details

These calculations can only occur if the periods of the cosinor are harmonic - as in, the longest period is a integer multiple of the smallest period (known as the fundamental frequency). Otherwise, these statistics are not accurate or interpretable.