Recurrent Event Summary Table by Group

recur_summary Creates a table with summary of recurrent events

recur_summary(data, covar)

Arguments

data
Recurrent event data in marginal format. There must be an ID column. Must merge in the covariate of interest into this data set.

covar
Name of covariate of interest to serve as grouping variable.

Value

Summary table by grouping variable, can be placed into a latex environment with kableExtra. Assumes that death events may be present when most recent non-EVENT has status 1.

Details

This function allows for taking the output of card::recur_survival_table marginal format repeat event data, and creates a summary table that describes the number of events by strata/event.

Examples

# setup
library(magrittr)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(kableExtra)

#> 
#> Attaching package: ‘kableExtra’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     group_rows

data("mims")

# Marginal data (as above)
marg <-
  recur_survival_table(
    mims, "patid", "first_visit_date_bl", "ldka",
    c("mi_date_1", "mi_date_2", "mi_date_3"), "marginal", "DEATH_CV_YN"
  )

#> Warning: `cols` is now required.
#> Please use `cols = c(data)`

# Grouping variable
covar <- "RDR_ISCHEMIA_M_YN_bl"
df <- left_join(marg, mims[c("patid", covar)], by = c("ID" = "patid"))

# Make summary table
tbl <- recur_summary(df, covar)

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `RDR_ISCHEMIA_M_YN_bl` contains implicit NA, consider using `forcats::fct_explicit_na`

tbl %>%
  kable("latex", caption = "Title", booktabs = TRUE) %>%
  kable_styling(font_size = 8)

#> \begin{table}
#> 
#> \caption{\label{tab:}Title}
#> \centering
#> \fontsize{8}{10}\selectfont
#> \begin{tabular}[t]{lrrrrrr}
#> \toprule
#> EVENT & Count\_RDR\_ISCHEMIA\_M\_YN\_bl\_0 & Count\_RDR\_ISCHEMIA\_M\_YN\_bl\_1 & Count\_RDR\_ISCHEMIA\_M\_YN\_bl\_NA & Time\_RDR\_ISCHEMIA\_M\_YN\_bl\_0 & Time\_RDR\_ISCHEMIA\_M\_YN\_bl\_1 & Time\_RDR\_ISCHEMIA\_M\_YN\_bl\_NA\\
#> \midrule
#> EVENT\_DATE\_0 & 250 & 50 & 7 & 1065.5320 & 1141.0600 & 1009.286\\
#> EVENT\_DATE\_1 & 26 & 10 & 2 & 669.3077 & 531.1000 & 337.500\\
#> EVENT\_DATE\_2 & 5 & 4 & 1 & 790.0000 & 536.2500 & 904.000\\
#> EVENT\_DATE\_3 & 1 & 1 & 1 & 281.0000 & 365.0000 & 986.000\\
#> EVENT\_DATE\_DEATH & 3 & 3 & NA & 598.0000 & 500.3333 & NA\\
#> \bottomrule
#> \end{tabular}
#> \end{table}